+++ src/Delizious.Toolkit/Match.cs
@@ -315,6 +315,41 @@
             return Match<T>.Create(CompareMatch<T>.GreaterThanOrEqualTo(reference, comparer));
         }
 
+        /// <summary>
+        /// Creates a <see cref="Match{T}"/> instance that matches successfully when the value to match is less than the specified <paramref name="reference"/> value.
+        /// </summary>
+        /// <typeparam name="T">
+        /// The type of the value to match.
+        /// </typeparam>
+        /// <param name="reference">
+        /// The instance a value to match must be less than to match successfully.
+        /// </param>
+        /// <param name="comparer">
+        /// The <see cref="IComparer{T}"/> instance to determine whether a value to match is less than the specified <paramref name="reference"/> value.
+        /// </param>
+        /// <returns>
+        /// A new <see cref="Match{T}"/> instance that matches successfully when the value to match is less than the specified <paramref name="reference"/> value.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        /// <para><paramref name="reference"/> is <c>null</c>.</para>
+        /// <para>- or -</para>
+        /// <para><paramref name="comparer"/> is <c>null</c>.</para>
+        /// </exception>
+        public static Match<T> LessThan<T>([NotNull] T reference, [NotNull] IComparer<T> comparer)
+        {
+            if (ReferenceEquals(reference, null))
+            {
+                throw new ArgumentNullException(nameof(reference));
+            }
+
+            if (ReferenceEquals(comparer, null))
+            {
+                throw new ArgumentNullException(nameof(comparer));
+            }
+
+            return Match<T>.Create(CompareMatch<T>.LessThan(reference, comparer));
+        }
+
         private sealed class CompareMatch<T> : IMatch<T>
         {
             private delegate bool Comparison(T left, T right, IComparer<T> comparer);
@@ -343,6 +378,12 @@
 
             private static bool GreaterThanOrEqualToComparison(T left, T right, IComparer<T> comparer)
                 => comparer.Compare(left, right) >= 0;
+
+            public static CompareMatch<T> LessThan(T reference, IComparer<T> comparer)
+                => new CompareMatch<T>(reference, comparer, LessThanComparison);
+
+            private static bool LessThanComparison<T>(T left, T right, IComparer<T> comparer)
+                => comparer.Compare(left, right) < 0;
 
             public bool Matches(T value)
                 => this.comparison(value, this.reference, this.comparer);