+++ CHANGELOG.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable match that matches when a value is not the same as a given reference value (#15)
 - Enable match that matches when a value is greater than a given reference value using a comparer (#17)
 - Enable match that matches when a value is greater than or equal to a given reference value using a comparer (#19)
+- Enable match that matches when a value is less than a given reference value using a comparer (#21)